move more stuff to global
