+++ src/locales/en/index.ts
@@ -22,7 +22,6 @@ import music from './music';
 import person from './person';
 import phone_number from './phone_number';
 import science from './science';
-import system from './system';
 import team from './team';
 import vehicle from './vehicle';
 import word from './word';
@@ -48,7 +47,6 @@ const en: LocaleDefinition = {
   person,
   phone_number,
   science,
-  system,
   team,
   vehicle,
   word,
