+++ src/locales/global/color/index.ts
@@ -3,10 +3,10 @@
  * Run 'pnpm run generate:locales' to update.
  */
 import type { ColorDefinitions } from '../../..';
-import human from './human';
+import space from './space';
 
 const color: ColorDefinitions = {
-  human,
+  space,
 };
 
 export default color;
