+++ src/locales/global/color/space.ts
