+++ src/locales/global/database/collation.ts
