+++ src/locales/global/database/type.ts
