+++ src/locales/global/hacker/abbreviation.ts
