+++ src/locales/global/hacker/index.ts
@@ -3,18 +3,10 @@
  * Run 'pnpm run generate:locales' to update.
  */
 import type { HackerDefinitions } from '../../..';
-import adjective from './adjective';
-import ingverb from './ingverb';
-import noun from './noun';
-import phrase from './phrase';
-import verb from './verb';
+import abbreviation from './abbreviation';
 
 const hacker: HackerDefinitions = {
-  adjective,
-  ingverb,
-  noun,
-  phrase,
-  verb,
+  abbreviation,
 };
 
 export default hacker;
