+++ src/locales/global/internet/http_status_code.ts
