+++ src/locales/global/internet/index.ts
@@ -4,9 +4,11 @@
  */
 import type { InternetDefinitions } from '../../..';
 import emoji from './emoji';
+import http_status_code from './http_status_code';
 
 const internet: InternetDefinitions = {
   emoji,
+  http_status_code,
 };
 
 export default internet;
