+++ src/locales/global/location/index.ts
@@ -1,0 +1,14 @@
+/*
+ * This file is automatically generated.
+ * Run 'pnpm run generate:locales' to update.
+ */
+import type { LocationDefinitions } from '../../..';
+import country_code from './country_code';
+import time_zone from './time_zone';
+
+const location: LocationDefinitions = {
+  country_code,
+  time_zone,
+};
+
+export default location;
