+++ src/locales/global/system/directoryPaths.ts
