+++ src/locales/global/system/index.ts
