+++ src/locales/global/system/mimeTypes.ts
